Add validation to 'rm'
The power of MOCK :D

+++ controller/core.py
@@ -2,7 +2,9 @@
 # -*- coding: utf-8 -*-
 
 import os
+import os.path
 
 
 def rm(filename):
-    os.remove(filename)
+    if os.path.isfile(filename):
+        os.remove(filename)
